navigation from sub component working
